Basic calculator opperations work! additionally, all doubles were changed to int's because only division could reach a floating point number

+++ Calculator_Back-End/Calculator.cs
@@ -5,25 +5,25 @@ namespace Calculator_Back_End
     public class Calculator
     {
 
-        public static double Add(double x, double y)
+        public static int Add(int x, int y)
         {
             //throw new NotImplementedException();
             return x + y;
         }
 
-        public static double Subtract(double x, double y)
+        public static int Subtract(int x, int y)
         {
             //throw new NotImplementedException();
             return x - y;
         }
 
-        public static double Multiply(double x, double y)
+        public static int Multiply(int x, int y)
         {
             //throw new NotImplementedException();
             return x * y;
         }
 
-        public static double Divide(double x, double y)
+        public static int Divide(int x, int y)
         {
             //throw new NotImplementedException();
             return x / y;
